timstamp return type from unsigned long to long long int

+++ helpers/general/include/opensbv/helpers/general/GeneralHelper.h
@@ -38,7 +38,7 @@ namespace opensbv {
                 static std::vector<std::string> explode(std::string const & s, char delim);
 
                 /// Get current timestamp
-                static unsigned long GetTimestamp();
+                static long long int GetTimestamp();
 
                 /// Get current datetime
                 static std::string GetDateTime();
